Util fixes

+++ utils/qrad3/trace.c
@@ -65,7 +65,7 @@ void MakeTnodes (dmodel_t *bm)
 {
 	// 32 byte align the structs
 	tnodes = malloc( (numnodes+1) * sizeof(tnode_t));
-	tnodes = (tnode_t *)(((int)tnodes + 31)&~31);
+	tnodes = (tnode_t *)(((intptr_t)tnodes + 31)&~31);
 	tnode_p = tnodes;
 
 	MakeTnode (0);
